btlp: seven units out of the middles of five asm blocks
Twenty-four routines come out of btlp_57, btlp_64, btlp_63, btlp_57_x1_d and
btlp_24_x2, nineteen of them exact and five behind NON_MATCHING. The BTLP
checksum is unchanged.

reach.c - the reach grid. D_800F4DC8 is g_btl_reach, five lanes by fifteen
rows, indexed [lane][row]: the enemies stand in rows 0..4 and the party ten
further on. Every marker clears it, lights a shape and marks whoever stands on
a lit cell. The two shape tables are named, and the three unnamed routines are
BtlMarkEnemiesAround, BtlMarkPartyAround and BtlPartyPickableMask. Five exact;
BtlPickAiTarget is one instruction out.

editboard.c - the debug character editor board: opened from seven parts, shut,
and filled from one fighter. All three exact, and its part table is emitted as
.data with bytes and relocations identical. The twenty-two digit cells and the
name buffer it writes are named, the two cells that were base-plus-offset are
separate symbols, and Char.pad1C and pad56 are really an int and a byte.

pickother.c - the second pair of random pickers, which charm uses. Both are
guarded at 98.66% on one delay-slot copy.

roundtick.c - BtlClearTalkMarks counted every fighter's timers down as a round
ends, so it is BtlCountDownRound. BtlActor.unkDE and the two timed conditions
are named. Guarded at 96.32%.

placecursor.c - the placement menu's previews: BtlStandPreset, BtlPlacePreset
and BtlMarkMovedMembers. D_800F47F8 is g_btl_formation_before, and layout 8 is
the live grid. BtlMarkMovedMembers is exact; the other two are guarded.

menunav.c - the five menus whose cursor walks a table of neighbours: the debug
board, the layout picker, the orders menu, the tactics page and the spell list.
All five exact, with their neighbour and spot tables as .data. BtlActor.unkC8
is the member's tactic.

stocklist.c - the negotiation board's loops and the Persona stock list on it.
Three exact; BtlBuildStockList is guarded at 95.66%.

Prototypes move into round.h, formation.h, menu.h, board.h, pick.h and input.h,
and the callers' local externs go. Seven levers that produced these matches are
added to Matching Nudges. The working-tree edits to clock.c, message.c and
pickable.c come along.

+++ src/btlp/pickable.c
@@ -68,7 +68,7 @@ int BtlAnyMemberTargetable(void)
 {
     /* Eight bytes of locals the routine reserves and never writes. */
     long unused[2];
-    int  status;
+    short status;
     int  i;
 
     i = 0;
